fix(introspect): decode composite field ordinal as i16
`pg_attribute.attnum` is a smallint, but `fetch_composite_types` decoded
it as `i32`. sqlx applies no implicit widening, so introspecting any
schema that declares a standalone composite type died with:

    ColumnDecode { index: "5", source: "mismatched types; Rust type
    `i32` (as SQL type `INT4`) is not compatible with SQL type `INT2`" }

The other ordinal reads in this module go through
`information_schema.columns.ordinal_position`, whose `cardinal_number`
domain really is an int4 — only the pg_catalog query was wrong.

Schemas with no standalone composite type return zero rows and never
decode anything, which is why this stayed hidden.

Also adds the `e2e_postgres` test target that the `test-postgres` CI job
already invokes but which never existed, so the job was silently passing
over nothing. It covers the composite regression plus basic table and
enum introspection.

Fixes #4

+++ crates/sqlx_gen/src/introspect/postgres.rs
@@ -431,7 +431,10 @@ async fn fetch_composite_types(
     pool: &PgPool,
     schemas: &[String],
 ) -> Result<Vec<CompositeTypeInfo>> {
-    let rows = sqlx::query_as::<_, (String, String, String, String, String, i32)>(
+    // `pg_attribute.attnum` is a smallint, unlike `information_schema`'s
+    // `ordinal_position` (int4) used elsewhere in this module. sqlx does no
+    // implicit widening, so this must be decoded as i16.
+    let rows = sqlx::query_as::<_, (String, String, String, String, String, i16)>(
         r#"
         SELECT
             n.nspname AS schema_name,
@@ -483,7 +486,7 @@ async fn fetch_composite_types(
             udt_schema: None,
             is_nullable: nullable == "YES",
             is_primary_key: false,
-            ordinal_position: ordinal,
+            ordinal_position: i32::from(ordinal),
             schema_name: schema,
             column_default: None,
         });
